DONE: Modified tags.services/tags.controller following the reviews

+++ backend/src/tags/tags.controller.ts
@@ -42,8 +42,13 @@ export class TagsController {
   }
 
   @Delete('/:id')
-  removeTag(@Res() res: Response, @Param('id') id: string) {
-    this.tagService.remove(parseInt(id));
-    res.json(null);
+  async removeTag(@Res() res: Response, @Param('id') id: string) {
+    if (id) {
+      await this.tagService.remove(parseInt(id));
+      res.json(null);
+    }
+    res.status(HttpStatus.INTERNAL_SERVER_ERROR).json({
+      error: 'Something went wrong',
+    });
   }
 }
